requirements.txt
opencv-python>=4.9.0
pytesseract>=0.3.10
PyYAML>=6.0.1
requests>=2.32.3
numpy>=1.26.0
